Update README.md
# publicawareness
A platform designed to expand financial aid, with a variety, diverse world of options, regardless of your credit score.*
The objective is to create software that will eliminate the human element out of the equation. 
The layers of human elements to operate a platform that allows any organization to take on the level of demand the particular trade is experiencing, scaling without compromise is almost invevitable.
Scaling in this manner, human elements also create certain levels of risk that can create monetary liabilities.
While equally significantly reducing the costs, for a service that is excessively being charged for. 
All while profit margins actually increase. We are redefining the laws of physics on a mumerical basis.
This will allow for more charitable donations at signifcant levels, towards other open collectives that are contributing in our way on another towards something that is giving back to our community, in any genuine, but meeting needs in demand. 
Currently the standard service fee nationwide, with the exception of a few states, due to states restrictions of fee limits, based on the overall debt of conversation. for someone to consolidate there debt with a debt relief or debt resolution like program. 
The recession we are in is real, very much so. A recession is defined as two consecutive quarters of negative economic growth. We surpassed, with the first few quarters this year. The data the federal reserve has changed its stance on where our country stands, strength and or weakness subjects. Financially speaking. 
Various resources have confirmed that consumers are at the highest peak of credit card usage, ever, in the history of the United States of America. Since the monetary system was put in place.
Consumers are now coming to a crossroads that is going to force them into a corner with limited options. 1) Bankruptcy 2) Credit Counseling 3) Debt Settlement 4) Debt Relief ) Debt Resolution.
Bankruptcy; everyone loses. Ch 7 Public Record 10 years - Ch 13 Public Record for 7 years| Credit Counselling; temporary solution that leaves a permanent mark on credit. Credit counseling gets reported to the credit bureaus, such like bankruptcies do.
Debt Settlement; too much risk involved. No guarantees of successfully completing plan. Debt Relief) Credit is compromised for the foreseeable future.
Now we arrive at our winner; Debt Resolution. Very similar to debt settlement and debt relief, but comes with tools that enable you to successful graduate the plan, restore your credit. Eliminating human elements, reduces labor costs and reduces risk.
With a software that is designed to operate as a complete "backend operating" debt resolution firm. This will significantly reduce the standard industry cost of 25% to 15%. Reducing the cost to the consumer by 15%, all while increasing the, font end operations margins by double by nearly double the amount. Its a win win for all involved.
A needs based product, that is seeing demand levels we have never seen before, are we're barely beginning to enter the boom of the industry is going to partake in.
The software will re-innovate and completely revamp the way others do business.
The ultimate objective of the software will be specifically designed to allow consumers to enroll themselves, as its human nature to not want to confront the issues your surrounded with everyday, all the way through signing the legally binding contract.
Providing a matter of convenience and a significantly reduces financial product to become debt free in the quickest time frame possible.